requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
attrs==19.3.0
Automat==0.8.0
bcrypt==4.1.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.3
certifi==2019.11.28
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
deluge==2.0.3
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==2.2.12
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
entrypoints==0.3
exceptiongroup==1.2.0
fast-api-gcs==0.0.11
fastapi==0.110.0
filelock==3.13.1
folder-color-common==0.0.88
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
hyperlink==19.0.0
identify==2.5.35
idna==3.6
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
motor==3.4.0
netifaces==0.10.4
nodeenv==1.8.0
oauthlib==3.1.0
olefile==0.46
orjson==3.10.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.11.0
pre-commit==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.5.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.16.2
pycups==1.9.73
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.6.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.1+zorin1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-libtorrent==1.1.13
python-multipart==0.0.9
pytz==2019.3
pyxdg==0.26
PyYAML==6.0.1
rencode==1.0.6
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.2.4
starlette==0.36.3
Twisted==18.9.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vboxapi==1.0
virtualenv==20.24.6
wadllib==1.3.3
watchfiles==0.21.0
websockets==12.0
xkit==0.0.0
zope.interface==4.7.1
zorin-appearance==4.1
zorin-connect==1.2
zorin-exec-guard==1.5
